requirements.txt
pandas
numpy
torch
transformers
sentence-transformers
faiss-cpu
fastapi
uvicorn[standard]
datasets

